backend/requirements.txt
# ─── Core API ─────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0

# ─── OCR ──────────────────────────────────────────────────────────────────────
transformers>=4.40.0
torch>=2.2.0
# torchvision is required by TrOCR processor
torchvision>=0.17.0
easyocr>=1.7.1
Pillow>=10.3.0

# ─── Image Pre-processing ─────────────────────────────────────────────────────
opencv-python-headless>=4.9.0
numpy>=1.26.0

# ─── Translation ──────────────────────────────────────────────────────────────
deep-translator>=1.11.4

# ─── TTS ──────────────────────────────────────────────────────────────────────
gTTS>=2.5.1

# ─── Fine-Tuning ──────────────────────────────────────────────────────────────
datasets>=2.19.0
evaluate>=0.4.1
jiwer>=3.0.3
accelerate>=0.30.0
# sentencepiece is required by some tokenizers
sentencepiece>=0.2.0
